profile page fix
There was an error in profile page,
Suppose, we have 6 moments in profile page, and once we click on profile
button on the top left of screen it shows 12 moments, that is , again 6
are added. In this way it adds 6 everytime we click.

I noticed this error, and fixed it.